+++ Delta.CertXplorer/src/Delta.CertXplorer.Common/Commanding/ICommand.cs
@@ -9,7 +9,8 @@ namespace Delta.CertXplorer.Commanding
         /// <summary>
         /// Runs the command with the specified arguments.
         /// </summary>
-        /// <param name="arguments">The arguments.</param>
-        void Run(params object[] arguments);
+        /// <param name="verb">The verb this command was invoked from (informative).</param>
+        /// <param name="arguments">The command arguments.</param>
+        void Run(IVerb verb, params object[] arguments);
     }
 }